Bullet class and more collitions
- Bullet class added.
- Bullet collition added.
- Enemy Health added.
- Enemy can die now.
- All deltatime references are now in float datatype.

+++ Juego2D/Ginfo.cpp
@@ -27,7 +27,7 @@ void Ginfo::Load()
 	}
 }
 
-void Ginfo::Update(const double &deltaTime)
+void Ginfo::Update(const float &deltaTime)
 {
 	timer += deltaTime;
 
@@ -36,7 +36,6 @@ void Ginfo::Update(const double &deltaTime)
 		text.setString("fps" + std::to_string((int)result) + " time: " + std::to_string(deltaTime) + "ms");
 		timer = 0;
 	}
-	
 }
 
 void Ginfo::Draw(sf::RenderWindow& window)
